associate metadataReport with registry
