+++ metadata/metadata_service.go
@@ -18,24 +18,29 @@
 package metadata
 
 import (
+	"fmt"
+	"strings"
+)
+
+import (
 	"github.com/dubbogo/gost/log/logger"
+	gxnet "github.com/dubbogo/gost/net"
+
+	perrors "github.com/pkg/errors"
 )
 
 import (
 	"dubbo.apache.org/dubbo-go/v3/common"
 	"dubbo.apache.org/dubbo-go/v3/common/constant"
+	"dubbo.apache.org/dubbo-go/v3/common/extension"
 	"dubbo.apache.org/dubbo-go/v3/metadata/info"
+	"dubbo.apache.org/dubbo-go/v3/protocol"
 )
 
 // version will be used by Version func
 const (
 	version              = "1.0.0"
 	allServiceInterfaces = "*"
-)
-
-var (
-	appMetadataInfoMap = make(map[string]*info.MetadataInfo)
-	metadataUrl        *common.URL
 )
 
 // MetadataService is used to define meta data related behaviors
@@ -52,10 +57,17 @@ type MetadataService interface {
 	GetMetadataInfo(revision string) (*info.MetadataInfo, error)
 	// GetMetadataServiceURL will return the url of metadata service
 	GetMetadataServiceURL() (*common.URL, error)
+	// SetMetadataServiceURL exporter to set url of metadata service, will not be exported by exporter,cause no error return
+	SetMetadataServiceURL(*common.URL)
 }
 
 // DefaultMetadataService is store and query the metadata info in memory when each service registry
 type DefaultMetadataService struct {
+	metadataUrl *common.URL
+}
+
+func (mts *DefaultMetadataService) SetMetadataServiceURL(url *common.URL) {
+	mts.metadataUrl = url
 }
 
 // GetExportedURLs get all exported urls
@@ -109,7 +121,7 @@ func (mts *DefaultMetadataService) Version() (string, error) {
 
 // GetMetadataServiceURL get url of MetadataService
 func (mts *DefaultMetadataService) GetMetadataServiceURL() (*common.URL, error) {
-	return metadataUrl, nil
+	return mts.metadataUrl, nil
 }
 
 func (mts *DefaultMetadataService) GetSubscribedURLs() ([]*common.URL, error) {
@@ -128,26 +140,55 @@ func (mts *DefaultMetadataService) MethodMapper() map[string]string {
 	}
 }
 
-func GetMetadataInfo(id string) *info.MetadataInfo {
-	return appMetadataInfoMap[id]
+// ServiceExporter is the ConfigurableMetadataServiceExporter which implement MetadataServiceExporter interface
+type ServiceExporter struct {
+	app, metadataType string
+	service           MetadataService
+	protocolExporter  protocol.Exporter
 }
 
-func AddService(registryId string, url *common.URL) {
-	if _, exist := appMetadataInfoMap[registryId]; !exist {
-		appMetadataInfoMap[registryId] = info.NewMetadataInfo(
-			url.GetParam(constant.ApplicationKey, ""),
-			url.GetParam(constant.ApplicationTagKey, ""),
-		)
+// Export will export the metadataService
+func (e *ServiceExporter) Export() error {
+	version, _ := e.service.Version()
+	tcp, err := gxnet.ListenOnTCPRandomPort("")
+	if err != nil {
+		panic(perrors.New(fmt.Sprintf("Get tcp port error, err is {%v}", err)))
 	}
-	appMetadataInfoMap[registryId].AddService(url)
+	err = tcp.Close()
+	if err != nil {
+		panic(perrors.New(fmt.Sprintf("Close tcp port error, err is {%v}", err)))
+	}
+	port := strings.Split(tcp.Addr().String(), ":")[1]
+	ivkURL := common.NewURLWithOptions(
+		common.WithPath(constant.MetadataServiceName),
+		common.WithProtocol(constant.DefaultProtocol),
+		common.WithPort(port),
+		common.WithParamsValue(constant.GroupKey, e.app),
+		common.WithParamsValue(constant.SerializationKey, constant.Hessian2Serialization),
+		common.WithParamsValue(constant.ReleaseKey, constant.Version),
+		common.WithParamsValue(constant.VersionKey, version),
+		common.WithParamsValue(constant.InterfaceKey, constant.MetadataServiceName),
+		common.WithParamsValue(constant.BeanNameKey, constant.SimpleMetadataServiceName),
+		common.WithParamsValue(constant.MetadataTypeKey, e.metadataType),
+		common.WithParamsValue(constant.SideKey, constant.SideProvider),
+	)
+	methods, err := common.ServiceMap.Register(ivkURL.Interface(), ivkURL.Protocol, ivkURL.Group(), ivkURL.Version(), e.service)
+	if err != nil {
+		formatErr := perrors.Errorf("The service %v needExport the protocol %v error! Error message is %v.",
+			ivkURL.Interface(), ivkURL.Protocol, err.Error())
+		logger.Errorf(formatErr.Error())
+		return formatErr
+	}
+	ivkURL.Methods = strings.Split(methods, ",")
+	proxyFactory := extension.GetProxyFactory("")
+	invoker := proxyFactory.GetInvoker(ivkURL)
+	e.protocolExporter = extension.GetProtocol(ivkURL.Protocol).Export(invoker)
+	e.service.SetMetadataServiceURL(ivkURL)
+	logger.Infof("[Metadata Service] The MetadataService exports urls : %v ", ivkURL)
+	return nil
 }
 
-func AddSubscribeURL(registryId string, url *common.URL) {
-	if _, exist := appMetadataInfoMap[registryId]; !exist {
-		appMetadataInfoMap[registryId] = info.NewMetadataInfo(
-			url.GetParam(constant.ApplicationKey, ""),
-			url.GetParam(constant.ApplicationTagKey, ""),
-		)
-	}
-	appMetadataInfoMap[registryId].AddSubscribeURL(url)
+// UnExport will unExport the metadataService
+func (e *ServiceExporter) UnExport() {
+	e.protocolExporter.UnExport()
 }
